services/platform/nvidia-llm-router/requirements.txt
# NVIDIA LLM Router Requirements

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
aiohttp==3.9.1
httpx==0.25.2

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1
aioredis==2.0.1

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0

# Utilities
backoff==2.2.1
tenacity==8.2.3

# Logging
structlog==23.2.0

# Testing (for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
